+++ apps/docs/content/components/OrderedList.mdx
@@ -1,8 +1,8 @@
 ---
-title: List
+title: Ordered List
 status: Experimental
-figma: 'https://www.figma.com/file/BJ95AjraesmRCWsKA013GS/Primer-Brand?node-id=354%3A9513'
-source: https://github.com/primer/brand/tree/main/src/List/OrderedList.tsx
+figma: 'https://www.figma.com/file/BJ95AjraesmRCWsKA013GS/branch/NzBMPIfPrFU9I87d3k7KVq/Primer-Brand?node-id=4045%3A17560&t=W8B9ArDngbzJCC2m-0'
+source: https://github.com/primer/brand/tree/main/src/List/OrderedList/OrderedList.tsx
 storybook: '/brand/storybook/?path=/story/components-orderedlist--default'
 description: Use the OrderedList component to render a list of ordered items.
 ---
