+++ apps/docs/content/components/UnorderedList.mdx
@@ -1,8 +1,8 @@
 ---
-title: List
+title: Unordered List
 status: Experimental
-figma: 'https://www.figma.com/file/BJ95AjraesmRCWsKA013GS/Primer-Brand?node-id=354%3A9513'
-source: https://github.com/primer/brand/tree/main/src/List/UnorderedList.tsx
+figma: 'https://www.figma.com/file/BJ95AjraesmRCWsKA013GS/branch/NzBMPIfPrFU9I87d3k7KVq/Primer-Brand?node-id=4045%3A17093&t=XJ9W58VgtbVo74gb-0'
+source: https://github.com/primer/brand/tree/main/src/List/UnorderedList/UnorderedList.tsx
 storybook: '/brand/storybook/?path=/story/components-unorderedlist--default'
 description: Use the UnorderedList component to render a list of unordered items.
 ---
@@ -29,7 +29,7 @@ import {UnorderedList} from '@primer/react-brand'
 </UnorderedList>
 ```
 
-### Checked UnorderedList
+### Checked variant
 
 ```jsx live
 <UnorderedList variant="checked">
@@ -52,7 +52,7 @@ import {UnorderedList} from '@primer/react-brand'
 | Name        | Type                                   |  Default  | Description                                                |
 | :---------- | :------------------------------------- | :-------: | :--------------------------------------------------------- |
 | `className` | `string`                               |           | Sets a custom class on the root element                    |
-| `variant`   | `default`, `checked`                   | `default` | Specify alternative List appearance                        |
+| `variant`   | `default`, `checked`                   | `default` | Specify alternative leading visual for list items          |
 | `children`  | `React.ReactNode`, `React.ReactNode[]` |           | Content to be displayed within the UnorderedList component |
 
 ### UnorderedList.Item
