--- packages/react/src/List/List.module.css
@@ -1,35 +1,0 @@
-.List {
-  display: flex;
-  flex-direction: column;
-  margin: 0;
-  list-style: none;
-  gap: var(--base-size-8);
-  padding-inline-start: 0px;
-  font-family: var(--brand-body-fontFamily);
-  font-weight: var(--brand-text-weight-400);
-  font-size: var(--brand-text-size-200);
-  line-height: var(--brand-text-lineHeight-400);
-  color: var(--brand-color-text-default);
-}
-
-ol.List {
-  counter-reset: li;
-}
-
-.ListItem {
-  display: flex;
-  gap: var(--base-size-12);
-  align-items: center;
-}
-
-.ListItem__leading-visual {
-  fill: var(--brand-color-text-muted);
-}
-
-ol.List > .ListItem::before {
-  content: counter(li) '.';
-  counter-increment: li;
-  min-width: var(--base-size-16);
-  text-align: center;
-  color: var(--brand-color-text-muted);
-}
