+++ packages/react/src/List/ListItem/ListItem.module.css
@@ -1,0 +1,17 @@
+.ListItem {
+  display: flex;
+  gap: var(--base-size-12);
+  align-items: center;
+}
+
+.ListItem__leading-visual {
+  fill: var(--brand-color-text-muted);
+}
+
+ol > .ListItem::before {
+  content: counter(li) '.';
+  counter-increment: li;
+  min-width: var(--base-size-16);
+  text-align: center;
+  color: var(--brand-color-text-muted);
+}
