+++ packages/react/src/List/ListItem/ListItem.module.css.d.ts
@@ -1,5 +1,4 @@
 declare const styles: {
-  readonly "List": string;
   readonly "ListItem": string;
   readonly "ListItem__leading-visual": string;
 };
