+++ packages/react/src/List/OrderedList/OrderedList.module.css
@@ -1,0 +1,3 @@
+ol.OrderedList {
+  counter-reset: li;
+}
