+++ packages/react/src/List/OrderedList/OrderedList.module.css.d.ts
@@ -1,0 +1,5 @@
+declare const styles: {
+  readonly "OrderedList": string;
+};
+export = styles;
+
