+++ packages/react/src/List/list-shared.module.css.d.ts
@@ -1,0 +1,5 @@
+declare const styles: {
+  readonly "List": string;
+};
+export = styles;
+
